+++ src/commands/api.ts
@@ -1,6 +1,4 @@
-import fetch from 'node-fetch';
 import * as vscode from 'vscode';
-import * as Commands from './index';
 import * as Utility from '../utility/index';
 import * as Service from '../service/index';
 import { apiWorkflows, getWorkflow } from './apiFlows';
@@ -8,8 +6,8 @@ import { apiWorkflows, getWorkflow } from './apiFlows';
 let disposable: vscode.Disposable;
 let disposableStatusBar: vscode.StatusBarItem;
 
-export function register(context: vscode.ExtensionContext) {
-    disposable = vscode.commands.registerCommand(Commands.shared.commandNames.api, async () => {
+async function register() {
+    disposable = vscode.commands.registerCommand(Service.command.commandNames.api, async () => {
         const endpoint = await Service.api.pick();
         if (!endpoint) {
             return;
@@ -45,20 +43,21 @@ export function register(context: vscode.ExtensionContext) {
 
     disposableStatusBar = vscode.window.createStatusBarItem(vscode.StatusBarAlignment.Left, 0);
     disposableStatusBar.text = 'ᕫ API Query';
-    disposableStatusBar.command = Commands.shared.commandNames.api;
+    disposableStatusBar.command = Service.command.commandNames.api;
     disposableStatusBar.tooltip = 'Query Data from the Blockchain';
     disposableStatusBar.show();
 
-    Commands.shared.installed.api = true;
+    const context = await Utility.context.get();
     context.subscriptions.push(disposable);
+    return () => {
+        if (disposable) {
+            disposable.dispose();
+        }
+
+        if (disposableStatusBar) {
+            disposableStatusBar.dispose();
+        }
+    };
 }
 
-export function dispose() {
-    if (disposable) {
-        disposable.dispose();
-    }
-
-    if (disposableStatusBar) {
-        disposableStatusBar.dispose();
-    }
-}
+Service.command.register('api', register);