+++ src/commands/createWallet.ts
@@ -1,12 +1,11 @@
 import * as vscode from 'vscode';
 import * as Service from '../service/index';
-import * as Commands from './index';
 import * as Utility from '../utility/index';
 
 let disposable: vscode.Disposable;
 
-export function register(context: vscode.ExtensionContext) {
-    disposable = vscode.commands.registerCommand(Commands.shared.commandNames.createWallet, async () => {
+async function register() {
+    disposable = vscode.commands.registerCommand(Service.command.commandNames.createWallet, async () => {
         if (Service.wallet.exists()) {
             vscode.window.showInformationMessage('Wallet already exists!');
             return;
@@ -31,14 +30,16 @@ export function register(context: vscode.ExtensionContext) {
         }
     });
 
-    Commands.shared.installed.createWallet = true;
+    const context = await Utility.context.get();
     context.subscriptions.push(disposable);
+
+    return () => {
+        if (!disposable) {
+            return;
+        }
+
+        disposable.dispose();
+    };
 }
 
-export function dispose() {
-    if (!disposable) {
-        return;
-    }
-
-    disposable.dispose();
-}
+Service.command.register('createWallet', register);