+++ src/service/index.ts
@@ -1,2 +1,3 @@
 export * as wallet from './wallet';
 export * as api from './api';
+export * as command from './command';