+++ src/utility/context.ts
@@ -1,4 +1,5 @@
 import * as vscode from 'vscode';
+import * as Utility from '../utility/index';
 
 let extContext: vscode.ExtensionContext;
 
@@ -6,6 +7,17 @@ export function init(context: vscode.ExtensionContext) {
     extContext = context;
 }
 
-export function get(): vscode.ExtensionContext {
+/**
+ * Returns the extension context after it is assigned
+ *
+ * @export
+ * @return {Promise<vscode.ExtensionContext>}
+ */
+export async function get(): Promise<vscode.ExtensionContext> {
+    if (typeof extContext === 'undefined') {
+        await Utility.sleep(500);
+        return get();
+    }
+
     return extContext;
 }