+++ src/utility/form.ts
@@ -1,7 +1,7 @@
 import * as Utility from './index';
 
-export function create<T = any>(title: string, fields: Array<string>) {
-    const formView = Utility.view.create(title, 'form.html');
+export async function create<T = any>(title: string, fields: Array<{ name: string; type: string }>) {
+    const formView = await Utility.view.create(title, 'form.html');
     formView.post({ event: 'setform', data: fields });
     return new Promise((resolve: (result: T | undefined) => void) => {
         let disposed = false;