+++ src/utility/state.ts
@@ -1,16 +1,19 @@
 import * as Utility from './index';
 
-export function set(key: string, value: any) {
+export async function set(key: string, value: any) {
     const keyName = `ultra.${key}`;
-    Utility.context.get().globalState.update(keyName, value);
+    const context = await Utility.context.get();
+    context.globalState.update(keyName, value);
 }
 
-export function get<T = unknown>(key: string): T | undefined {
+export async function get<T = unknown>(key: string): Promise<T | undefined> {
     const keyName = `ultra.${key}`;
-    return Utility.context.get().globalState.get(keyName);
+    const context = await Utility.context.get();
+    return context.globalState.get(keyName);
 }
 
-export function has(key: string): boolean {
+export async function has(key: string): Promise<boolean> {
     const keyName = `ultra.${key}`;
-    return typeof Utility.context.get().globalState.get(keyName) !== 'undefined';
+    const context = await Utility.context.get();
+    return typeof context.globalState.get(keyName) !== 'undefined';
 }